+++ UncomplicatedCustomRoles/API/Features/CustomModules/AmnesiaResistance.cs
@@ -1,0 +1,15 @@
+﻿/*
+ * This file is a part of the UncomplicatedCustomRoles project.
+ * 
+ * Copyright (c) 2023-present FoxWorn3365 (Federico Cosma) <me@fcosma.it>
+ * 
+ * This file is licensed under the GNU Affero General Public License v3.0.
+ * You should have received a copy of the AGPL license along with this file.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace UncomplicatedCustomRoles.API.Features.CustomModules
+{
+    internal class AmnesiaResistance : CustomModule
+    { }
+}
